Fix manual test
If we manually delete the key right after setting a value, and
we do not let it iterate to receive a change, the change we are
getting is the one from setting the value and not the one from
manually deleting the key. Split it so we iterate the main loop
twice to get 2 signals.